move error handling to service

+++ paintings_app_backend/src/main/java/com/example/paintings_app_backend/service/PaintingService.java
@@ -4,11 +4,12 @@ import com.example.paintings_app_backend.domain.Painting;
 import com.example.paintings_app_backend.repository.IPaintingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Transactional
@@ -28,22 +29,29 @@ public class PaintingService {
         }
     }
 
-    public Optional<Painting> getById(int id) {
-        return repository.findById(id);
+    public Painting getById(int id) {
+        var result = repository.findById(id);
+        if (result.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Painting not found");
+        return result.get();
     }
 
     public Painting add(Painting painting) {
         if (painting.getId() != 0)
-            throw new IllegalArgumentException("Id must be 0 for new entity");
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Id must be 0 for new entity");
         return repository.save(painting);
     }
 
     public Painting update(int id, Painting painting) {
+        if (repository.findById(id).isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Painting not found");
         painting.setId(id);
         return repository.save(painting);
     }
 
     public void delete(int id) {
+        if (repository.findById(id).isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Painting not found");
         repository.deleteById(id);
     }
 }
